requirements.txt
mkdocs==1.6.0
mkdocs-material==9.5.32
mkdocs-static-i18n==1.3.0
mkdocs-gen-files==0.5.0
mkdocs-macros-plugin==1.3.9
azure-storage-blob==12.19.1